[front] clean - replace interface by type - tidy code in new files
