+++ frontend/src/components/map/const.ts
@@ -1,0 +1,8 @@
+export const pathMapper: Record<string, string> = {
+  Railway: "chart.paths.railway",
+  Road: "chart.paths.road",
+  Bike: "chart.paths.bike",
+  Flight: "chart.paths.flight",
+  Ferry: "chart.paths.ferry",
+  Sail: "chart.paths.sail",
+};
